requirements.txt
matplotlib==3.9.2
numpy==2.1.2
paho_mqtt==2.1.0
pandas==2.2.3
python-dotenv==1.0.1
torch==2.5.0
